fixed updateDom

+++ src/js/utils.js
@@ -58,7 +58,7 @@ utils.numberToHour = function(number){
 utils.hourToNumber = function(hour){
   const parts = hour.split(':');
 
-  return parseInt(parts[0]) + parseInt(parts[1])/60;
+  return parseInt(parts[0]) + parseFloat(parts[1])/60;
 };
 
 utils.dateToStr = function(dateObj){
